requirements.txt
aiogram>=3.5.0 # Используем aiogram 3+
google-generativeai>=0.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
openpyxl>=3.1.0 # Для поддержки .xlsx в pandas
PyPDF2>=3.0.0
pytesseract>=0.3.10
# Pillow необходим для pytesseract и Gemini Vision
Pillow>=10.0.0
# SpeechRecognition требует установки PortAudio (sudo apt install portaudio19-dev)
SpeechRecognition>=3.10.0
# pydub требует установки ffmpeg (sudo apt install ffmpeg)
pydub>=0.25.0
loguru>=0.7.0
aiosqlite>=0.19.0
gTTS>=2.5.0
# googletrans может быть нестабилен, используйте конкретную версию или форк при необходимости
# googletrans==4.0.0rc1
googletrans>4.0.0 # Попробуем последнюю доступную